requirements.txt
aiogram==3.23.0
python-dotenv==0.21.0